requirements.txt

fastapi
uvicorn[standard]
langchain
langchain-community
transformers
torch
sentence-transformers
faiss-cpu
pypdf

unstructured[local-inference] # Glavni paket za obradu fajlova
python-docx                 # Za .docx fajlove
python-pptx                 # Za .pptx fajlove
# Za Windows, ovo je neophodno za detekciju tipa fajla
python-magic-bin

langchain-google-genai
python-dotenv

# Za OpenAI
openai
# Za Google Gemini
google-generativeai
# Za Ollama
ollama